ai_agents/central_agent_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
gunicorn
langchain
langgraph
langchain-groq